nomad v0.7.0 Mix.Tasks.Nomad.DatabaseInstance.Create

Task for automatically creating a remote SQL database instance on a pre-determined cloud provider. The instance creation is done through the cloud providers’ API’s (mainly Amazon RDS and Google Cloud SQL) and restricted to the common options and operations that are offered in those API’s.

More in depth creation of instances with other options and configurations not provided in here must be done in the cloud provider’s console or by other means.

Functions

run(args)

Specs

run(list) :: binary

Runs the task for the chosen cloud provider. The shell prompts and necessary input parameters change with the chosen provider.