nomad v0.5.1 Mix.Tasks.Nomad.Deploy

Task for automatically deploying production releases of Elixir and Phoenix applications to remote hosts. This task is mainly targeted at cloud hosts like Amazon EC2 instances or Google Compute Engine instances, but will work with any UNIX remote machine.

The deployment is done through SSH and it follows the officials guidelines at www.phoenixframework.org.

Functions

run(args)

Deploy the application in the current project’s directory. Arguments must be in order: : Required - the remote host hostname or IP address
: Required - the port for the production application : Required - the SSH Key for accessing the remote host, only the filename and it must be in ~/.ssh : Required - the username to use to access the remote host <-f> : Optional - bypass all the questions during deployment