nomad v0.5.1 Nomad.AWS.SQL.Helper

Consult the documentation in ExAws.EC2.Client

Functions

attach_volume(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.attach_volume/3.

attach_volume(arg0, arg1, arg2, arg3)

Callback implementation for c:ExAws.EC2.Client.attach_volume/4.

authorize_security_group_egress(arg0)

Callback implementation for c:ExAws.EC2.Client.authorize_security_group_egress/1.

authorize_security_group_egress(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.authorize_security_group_egress/2.

authorize_security_group_ingress()

Callback implementation for c:ExAws.EC2.Client.authorize_security_group_ingress/0.

authorize_security_group_ingress(arg0)

Callback implementation for c:ExAws.EC2.Client.authorize_security_group_ingress/1.

bundle_instance(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.bundle_instance/2.

bundle_instance(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.bundle_instance/3.

cancel_bundle_task(arg0)

Callback implementation for c:ExAws.EC2.Client.cancel_bundle_task/1.

cancel_bundle_task(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.cancel_bundle_task/2.

copy_image(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.copy_image/3.

copy_image(arg0, arg1, arg2, arg3)

Callback implementation for c:ExAws.EC2.Client.copy_image/4.

copy_snapshot(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.copy_snapshot/2.

copy_snapshot(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.copy_snapshot/3.

create_image(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_image/2.

create_image(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.create_image/3.

create_key_pair(arg0)

Callback implementation for c:ExAws.EC2.Client.create_key_pair/1.

create_key_pair(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_key_pair/2.

create_security_group(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_security_group/2.

create_security_group(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.create_security_group/3.

create_sg_with_local_public_ip_allowed(instance, engine)

Specs

create_sg_with_local_public_ip_allowed(instance :: binary, engine :: binary) ::
  :ok |
  :error

Create a new Security Group that enables inbound traffic from the current local machine’s public IP address. The Security Group will have the name ‘secgroup-’ and the port used will be determined by the provided SQL ‘engine’.

create_snapshot(arg0)

Callback implementation for c:ExAws.EC2.Client.create_snapshot/1.

create_snapshot(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_snapshot/2.

create_subnet(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_subnet/2.

create_subnet(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.create_subnet/3.

create_tags(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_tags/2.

create_tags(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.create_tags/3.

create_volume(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_volume/2.

create_volume(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.create_volume/3.

create_vpc(arg0)

Callback implementation for c:ExAws.EC2.Client.create_vpc/1.

create_vpc(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.create_vpc/2.

delete_key_pair(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_key_pair/1.

delete_key_pair(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_key_pair/2.

delete_snapshot(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_snapshot/1.

delete_snapshot(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_snapshot/2.

delete_subnet(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_subnet/1.

delete_subnet(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_subnet/2.

delete_tags(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_tags/1.

delete_tags(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_tags/2.

delete_volume(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_volume/1.

delete_volume(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_volume/2.

delete_vpc(arg0)

Callback implementation for c:ExAws.EC2.Client.delete_vpc/1.

delete_vpc(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.delete_vpc/2.

deregister_image(arg0)

Callback implementation for c:ExAws.EC2.Client.deregister_image/1.

deregister_image(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.deregister_image/2.

describe_account_attributes()

Callback implementation for c:ExAws.EC2.Client.describe_account_attributes/0.

describe_account_attributes(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_account_attributes/1.

describe_availability_zones()

Callback implementation for c:ExAws.EC2.Client.describe_availability_zones/0.

describe_availability_zones(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_availability_zones/1.

describe_bundle_tasks()

Callback implementation for c:ExAws.EC2.Client.describe_bundle_tasks/0.

describe_bundle_tasks(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_bundle_tasks/1.

describe_id_format()

Callback implementation for c:ExAws.EC2.Client.describe_id_format/0.

describe_id_format(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_id_format/1.

describe_image_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.describe_image_attribute/2.

describe_image_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.describe_image_attribute/3.

describe_images()

Callback implementation for c:ExAws.EC2.Client.describe_images/0.

describe_images(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_images/1.

describe_instance_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.describe_instance_attribute/2.

describe_instance_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.describe_instance_attribute/3.

describe_instance_status()

Callback implementation for c:ExAws.EC2.Client.describe_instance_status/0.

describe_instance_status(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_instance_status/1.

describe_instances()

Callback implementation for c:ExAws.EC2.Client.describe_instances/0.

describe_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_instances/1.

describe_key_pairs()

Callback implementation for c:ExAws.EC2.Client.describe_key_pairs/0.

describe_key_pairs(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_key_pairs/1.

describe_regions()

Callback implementation for c:ExAws.EC2.Client.describe_regions/0.

describe_regions(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_regions/1.

describe_security_groups()

Callback implementation for c:ExAws.EC2.Client.describe_security_groups/0.

describe_security_groups(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_security_groups/1.

describe_snapshot_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.describe_snapshot_attribute/2.

describe_snapshot_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.describe_snapshot_attribute/3.

describe_snapshots()

Callback implementation for c:ExAws.EC2.Client.describe_snapshots/0.

describe_snapshots(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_snapshots/1.

describe_subnets()

Callback implementation for c:ExAws.EC2.Client.describe_subnets/0.

describe_subnets(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_subnets/1.

describe_tags()

Callback implementation for c:ExAws.EC2.Client.describe_tags/0.

describe_tags(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_tags/1.

describe_volume_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_volume_attribute/1.

describe_volume_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.describe_volume_attribute/2.

describe_volume_status()

Callback implementation for c:ExAws.EC2.Client.describe_volume_status/0.

describe_volume_status(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_volume_status/1.

describe_volumes()

Callback implementation for c:ExAws.EC2.Client.describe_volumes/0.

describe_volumes(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_volumes/1.

describe_vpc_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.describe_vpc_attribute/2.

describe_vpc_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.describe_vpc_attribute/3.

describe_vpcs()

Callback implementation for c:ExAws.EC2.Client.describe_vpcs/0.

describe_vpcs(arg0)

Callback implementation for c:ExAws.EC2.Client.describe_vpcs/1.

detach_volume(arg0)

Callback implementation for c:ExAws.EC2.Client.detach_volume/1.

detach_volume(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.detach_volume/2.

enable_volume_io(arg0)

Callback implementation for c:ExAws.EC2.Client.enable_volume_io/1.

enable_volume_io(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.enable_volume_io/2.

get_console_output(arg0)

Callback implementation for c:ExAws.EC2.Client.get_console_output/1.

get_console_output(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.get_console_output/2.

get_password_data(arg0)

Callback implementation for c:ExAws.EC2.Client.get_password_data/1.

get_password_data(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.get_password_data/2.

import_key_pair(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.import_key_pair/2.

import_key_pair(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.import_key_pair/3.

modify_id_format(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_id_format/2.

modify_image_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_image_attribute/1.

modify_image_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_image_attribute/2.

modify_instance_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_instance_attribute/1.

modify_instance_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_instance_attribute/2.

modify_snapshot_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_snapshot_attribute/1.

modify_snapshot_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_snapshot_attribute/2.

modify_subnet_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_subnet_attribute/1.

modify_subnet_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_subnet_attribute/2.

modify_volume_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_volume_attribute/1.

modify_volume_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_volume_attribute/2.

modify_vpc_attribute(arg0)

Callback implementation for c:ExAws.EC2.Client.modify_vpc_attribute/1.

modify_vpc_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.modify_vpc_attribute/2.

monitor_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.monitor_instances/1.

monitor_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.monitor_instances/2.

new(opts \\ [])

reboot_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.reboot_instances/1.

reboot_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.reboot_instances/2.

register_image(arg0)

Callback implementation for c:ExAws.EC2.Client.register_image/1.

register_image(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.register_image/2.

report_instance_status(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.report_instance_status/2.

report_instance_status(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.report_instance_status/3.

reset_image_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.reset_image_attribute/2.

reset_image_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.reset_image_attribute/3.

reset_instance_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.reset_instance_attribute/2.

reset_instance_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.reset_instance_attribute/3.

reset_snapshot_attribute(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.reset_snapshot_attribute/2.

reset_snapshot_attribute(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.reset_snapshot_attribute/3.

revoke_security_group_egress(arg0)

Callback implementation for c:ExAws.EC2.Client.revoke_security_group_egress/1.

revoke_security_group_egress(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.revoke_security_group_egress/2.

revoke_security_group_ingress()

Callback implementation for c:ExAws.EC2.Client.revoke_security_group_ingress/0.

revoke_security_group_ingress(arg0)

Callback implementation for c:ExAws.EC2.Client.revoke_security_group_ingress/1.

run_instances(arg0, arg1, arg2)

Callback implementation for c:ExAws.EC2.Client.run_instances/3.

run_instances(arg0, arg1, arg2, arg3)

Callback implementation for c:ExAws.EC2.Client.run_instances/4.

start_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.start_instances/1.

start_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.start_instances/2.

stop_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.stop_instances/1.

stop_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.stop_instances/2.

terminate_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.terminate_instances/1.

terminate_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.terminate_instances/2.

unmonitor_instances(arg0)

Callback implementation for c:ExAws.EC2.Client.unmonitor_instances/1.

unmonitor_instances(arg0, arg1)

Callback implementation for c:ExAws.EC2.Client.unmonitor_instances/2.