nomad v0.7.0 Nomad.Storage

Interface for the Nomad Storage API. All the functions here will call their respective callbacks to the desired cloud’s client adapter.

Functions

create_storage(bucket, region, class, fun \\ &create_bucket/3)

create_storage!(bucket, region, class, fun \\ &create_bucket/3)

delete_item(bucket, object, fun \\ &delete_object/2)

delete_item!(bucket, object, fun \\ &delete_object/2)

delete_storage(bucket, fun \\ &delete_bucket/1)

delete_storage!(bucket, fun \\ &delete_bucket/1)

get_item(bucket, object, fun \\ &get_object/2)

get_item!(bucket, object, fun \\ &get_object/2)

get_item_acl(bucket, object, fun \\ &get_object_acl/2)

get_item_acl!(bucket, object, fun \\ &get_object_acl/2)

get_storage_acl(bucket, fun \\ &get_bucket_acl/1)

get_storage_acl!(bucket, fun \\ &get_bucket_acl/1)

get_storage_class(bucket, fun \\ &get_bucket_class/1)

get_storage_class!(bucket, fun \\ &get_bucket_class/1)

get_storage_region(bucket, fun \\ &get_bucket_region/1)

get_storage_region!(bucket, fun \\ &get_bucket_region/1)

list_classes()

Specs

list_classes :: [binary]

Lists all the available classes for Cloud Storage buckets.

list_items(bucket, fun \\ &list_objects/1)

list_items!(bucket, fun \\ &list_objects/1)

list_storages(fun \\ &list_buckets/0)

list_storages!(fun \\ &list_buckets/0)

put_item(bucket, filepath, storage_path, fun \\ &put_object/3)

put_item!(bucket, filepath, storage_path, fun \\ &put_object/3)